requirements.txt
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
attrs==25.3.0
certifi==2025.7.14
cffi==2.0.0
click==8.2.1
coloredlogs==15.0.1
cryptography==46.0.1
dnspython==2.7.0
dotenv==0.9.9
email_validator==2.2.0
Faker==37.4.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.4
fificore==0.0.29
greenlet==3.2.2
gunicorn==23.0.0
h11==0.16.0
hiredis==3.2.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
humanfriendly==10.0
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.8.0
orjson==3.11.1
outcome==1.3.0.post0
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.10
pycparser==2.23
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.1
pytest-asyncio==1.0.0
python-dotenv==1.1.0
python-multipart==0.0.20
python-ulid==1.1.0
PyYAML==6.0.2
redis==5.3.1
redis-om==0.3.5
rich==14.0.0
rich-toolkit==0.14.8
rignore==0.6.4
sentry-sdk==2.33.1
setuptools==80.9.0
shellingham==1.5.4
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.41
starlette==0.47.2
trio==0.30.0
typer==0.16.0
types-cffi==1.17.0.20250915
types-pyOpenSSL==24.1.0.20240722
types-redis==4.6.0.20241004
types-setuptools==80.9.0.20250822
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
